feat: rename useCachedFilters to useQueryCache
