+++ ui/src/composables/index.js
@@ -1,6 +1,5 @@
 export { default as useContext } from './use-context.js'
-export { default as useHistory } from './use-history.js'
 export { default as useForm } from './use-form.js'
+export { default as useHistory } from './use-history.js'
+export { default as useQueryCache } from './use-query-cache.js'
 export { default as useScreen } from './use-screen.js'
-
-export * as useCachedFilter from './use-cached-filters.js'
